+++ app/Events/NewNotificationEvent.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace App\Events;
+
+use App\Models\AppNotification;
+use Illuminate\Broadcasting\InteractsWithSockets;
+use Illuminate\Broadcasting\PrivateChannel;
+use Illuminate\Contracts\Broadcasting\ShouldBroadcastNow;
+use Illuminate\Foundation\Events\Dispatchable;
+use Illuminate\Queue\SerializesModels;
+
+class NewNotificationEvent implements ShouldBroadcastNow
+{
+    use Dispatchable, InteractsWithSockets, SerializesModels;
+
+    public $notification;
+
+    /**
+     * Create a new event instance.
+     */
+    public function __construct(AppNotification $notification)
+    {
+        $this->notification = $notification;
+    }
+
+    /**
+     * Get the channels the event should broadcast on.
+     *
+     * @return array<int, \Illuminate\Broadcasting\Channel>
+     */
+    public function broadcastOn(): array
+    {
+        return [
+            new PrivateChannel('App.Models.User.' . $this->notification->user_id),
+        ];
+    }
+    
+    /**
+     * The event's broadcast name.
+     */
+    public function broadcastAs(): string
+    {
+        return 'notification.new';
+    }
+}